+++ .github/workflows/pytest.yml
@@ -23,4 +23,4 @@ jobs:
         poetry install
     - name: Test with pytest
       run: |
-        poetry run pytest
+        poetry run pytest src tests
