+++ pyproject.toml
@@ -14,6 +14,10 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Astronomy",
 ]
 
+[tool.coverage.paths]
+source = ["src"]
+
+
 [tool.poetry.dependencies]
 python = "^3.7"
 numpy = "^1.19.4"
