+++ tests/data/tpf_test_03.fits
