+++ tests/data/tpf_test_06.fits
