+++ tests/data/tpf_test_08.fits
